// packages/aip/src/lib/gen/bearclaw/aip/v1/pagination_pb.ts
// @generated by protoc-gen-es v2.2.5 with parameter "target=ts"
// @generated from file bearclaw/aip/v1/pagination.proto (package bearclaw.aip.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file bearclaw/aip/v1/pagination.proto.
 */
export const file_bearclaw_aip_v1_pagination: GenFile = /*@__PURE__*/
  fileDesc("CiBiZWFyY2xhdy9haXAvdjEvcGFnaW5hdGlvbi5wcm90bxIPYmVhcmNsYXcuYWlwLnYxIlwKFVRlc3RQYWdpbmF0aW9uUmVxdWVzdBIOCgZwYXJlbnQYASABKAkSEQoJcGFnZV9zaXplGAIgASgFEhIKCnBhZ2VfdG9rZW4YAyABKAkSDAoEc2tpcBgEIAEoBWIGcHJvdG8z");

/**
 * @generated from message bearclaw.aip.v1.TestPaginationRequest
 */
export type TestPaginationRequest = Message<"bearclaw.aip.v1.TestPaginationRequest"> & {
  /**
   * @generated from field: string parent = 1;
   */
  parent: string;

  /**
   * @generated from field: int32 page_size = 2;
   */
  pageSize: number;

  /**
   * @generated from field: string page_token = 3;
   */
  pageToken: string;

  /**
   * @generated from field: int32 skip = 4;
   */
  skip: number;
};

/**
 * Describes the message bearclaw.aip.v1.TestPaginationRequest.
 * Use `create(TestPaginationRequestSchema)` to create a new message.
 */
export const TestPaginationRequestSchema: GenMessage<TestPaginationRequest> = /*@__PURE__*/
  messageDesc(file_bearclaw_aip_v1_pagination, 0);